Split extract post proc from libDensity to libPostpro (as in master) + clean print.
